backend/requirement.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# File Upload (Cloudinary)
cloudinary==1.38.0

# Payment Integration (PayHere - Sri Lanka)
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0